Start a little refactoring to make the Python packages more consistent
