+++ src/hermes/__main__.py
@@ -4,8 +4,8 @@
 
 # SPDX-FileContributor: Michael Meinel
 
-from hermes.cli import main
+from hermes.commands.cli import main
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     main()
